+++ src/sys/os/linux/mod.rs
@@ -117,13 +117,13 @@ crate::items! { // structural access: _mods, _pub_mods, _all
     pub use _pub_mods::*;
 
     mod _mods {
-        pub use super::namespace::*;
+        pub use super::{error::*, namespace::*};
         #[cfg(all(feature = "unsafe_syscall", not(miri)))]
         pub use super::terminal::*;
     }
     mod _pub_mods { #![allow(unused)]
         pub use super::{
-            consts::_all::*, error::*, io::_all::*, process::_all::*, thread::_all::*,
+            consts::_all::*, io::_all::*, process::_all::*, thread::_all::*,
         };
     }
     pub(super) mod _all { #![allow(unused)]